Split virtual root helpers and harden leaf rename handling
Move the virtual directory configuration, normalization, and virtual-root helper methods out of FileSystem.go into virtual_config.go.

Make DirINode embed baseDirNode so real and virtual directory nodes share the same cached child and negative-cache implementation.

Tighten virtual mutation checks so configured leaf paths stay visible for lookup and readdir but cannot themselves be renamed or replaced; only descendants are mutable.

Add a regression test covering leaf visibility versus mutation permissions.
