+++ internal/hopsfsmount/Dir.go
@@ -8,9 +8,7 @@ import (
 	"github.com/colinmarc/hdfs/v2"
 	"os"
 	"path"
-	"sync"
 	"syscall"
-	"time"
 
 	"bazil.org/fuse"
 	"bazil.org/fuse/fs"
@@ -20,13 +18,7 @@ import (
 
 // DirINode represents a real backend-backed directory node.
 type DirINode struct {
-	FileSystem    *FileSystem        // Pointer to the owning filesystem
-	Attrs         Attrs              // Cached attributes of the directory, TODO: add TTL
-	Parent        Pather             // Pointer to the parent directory (allows computing fully-qualified paths on demand)
-	children      map[string]fs.Node // Cached directory entries
-	negativeCache map[string]time.Time
-	childrenMutex sync.Mutex
-	dirMutex      sync.Mutex
+	baseDirNode
 }
 
 // Verify that *Dir implements necesary FUSE interfaces
@@ -43,20 +35,7 @@ var _ fs.NodeLinker = (*DirINode)(nil)
 var _ fs.NodeCreater = (*DirINode)(nil)
 var _ fs.NodeFsyncer = (*DirINode)(nil)
 
-// Returns absolute path of the dir in HDFS namespace
-func (dir *DirINode) AbsolutePath() string {
-	if dir.Parent == nil {
-		return dir.FileSystem.SrcDir
-	}
-	return path.Join(dir.Parent.AbsolutePath(), dir.Attrs.Name)
-}
-
-// Returns absolute path of the child item of this directory
-func (dir *DirINode) AbsolutePathForChild(name string) string {
-	return path.Join(dir.AbsolutePath(), name)
-}
-
-// Responds on FUSE request to get directory attributes
+// Responds on FUSE request to get directory attributes.
 func (dir *DirINode) Attr(ctx context.Context, a *fuse.Attr) error {
 	dir.lockMutex()
 	defer dir.unlockMutex()
@@ -73,96 +52,7 @@ func (dir *DirINode) Attr(ctx context.Context, a *fuse.Attr) error {
 	return dir.Attrs.ConvertAttrToFuse(a)
 }
 
-func (dir *DirINode) getChildInode(operation, name string) fs.Node {
-	dir.lockChildrenMutex()
-	defer dir.unlockChildrenMutex()
-
-	if dir.children == nil {
-		dir.children = make(map[string]fs.Node)
-		return nil
-	}
-
-	node := dir.children[name]
-	if node != nil {
-		logger.Debug("Children's List. getChildInode ", logger.Fields{Operation: operation, Parent: dir.AbsolutePath(), Child: name, NumChildren: len(dir.children)})
-	} else {
-		logger.Debug("Children's List. getChildInode. Not Found  ", logger.Fields{Operation: operation, Parent: dir.AbsolutePath(), Child: name, NumChildren: len(dir.children)})
-	}
-
-	return node
-}
-
-func (dir *DirINode) addOrUpdateChildInodeAttrs(operation, name string, attrs Attrs) fs.Node {
-	dir.lockChildrenMutex()
-	defer dir.unlockChildrenMutex()
-
-	if dir.children == nil {
-		dir.children = make(map[string]fs.Node)
-	}
-
-	shouldBeDir := (attrs.Mode & os.ModeDir) != 0
-	if node, ok := dir.children[name]; ok {
-		if shouldBeDir {
-			if dnode, ok := (node).(*DirINode); ok {
-				dnode.Attrs = attrs
-				logger.Debug("Children's List. addOrUpdateChildInodeAttrs. Update ", logger.Fields{Operation: operation, Parent: dir.AbsolutePath(), Child: name, NumChildren: len(dir.children)})
-				return node
-			}
-			node = &DirINode{FileSystem: dir.FileSystem, Parent: dir, Attrs: attrs}
-		} else {
-			if fnode, ok := (node).(*FileINode); ok {
-				fnode.Attrs = attrs
-				logger.Debug("Children's List. addOrUpdateChildInodeAttrs. Update ", logger.Fields{Operation: operation, Parent: dir.AbsolutePath(), Child: name, NumChildren: len(dir.children)})
-				return node
-			}
-			node = &FileINode{FileSystem: dir.FileSystem, Parent: dir, Attrs: attrs}
-		}
-
-		dir.children[name] = node
-		logger.Debug("Children's List. addOrUpdateChildInodeAttrs. Replace ", logger.Fields{Operation: operation, Parent: dir.AbsolutePath(), Child: name, NumChildren: len(dir.children)})
-		return node
-	} else {
-		var node fs.Node
-		if shouldBeDir {
-			node = &DirINode{FileSystem: dir.FileSystem, Parent: dir, Attrs: attrs}
-		} else {
-			node = &FileINode{FileSystem: dir.FileSystem, Parent: dir, Attrs: attrs}
-		}
-		dir.children[name] = node
-		logger.Debug("Children's List. addOrUpdateChildInodeAttrs. Add ", logger.Fields{Operation: operation, Parent: dir.AbsolutePath(), Child: name, NumChildren: len(dir.children)})
-		return node
-	}
-}
-
-func (dir *DirINode) removeChildInode(operation, name string) {
-	dir.lockChildrenMutex()
-	defer dir.unlockChildrenMutex()
-
-	if dir.children != nil {
-		delete(dir.children, name)
-		logger.Debug("Children's List. removeChildInode ", logger.Fields{Operation: operation, Parent: dir.AbsolutePath(), Child: name, NumChildren: len(dir.children)})
-	}
-}
-
-// used in rename. when an inode is moved from one dir to another
-func (dir *DirINode) adoptChildInode(operation, name string, node fs.Node) {
-	dir.lockChildrenMutex()
-	defer dir.unlockChildrenMutex()
-
-	if dir.children == nil {
-		dir.children = make(map[string]fs.Node)
-	}
-
-	if _, ok := dir.children[name]; ok {
-		logger.Debug("Children's List. Adopted inode. Replaced existing node ", logger.Fields{Operation: operation, Parent: dir.AbsolutePath(), Child: name, NumChildren: len(dir.children)})
-	} else {
-		logger.Debug("Children's List. Adopted inode. Added new node ", logger.Fields{Operation: operation, Parent: dir.AbsolutePath(), Child: name, NumChildren: len(dir.children)})
-	}
-
-	dir.children[name] = node
-}
-
-// Responds on FUSE request to lookup the directory
+// Responds on FUSE request to lookup the directory.
 func (dir *DirINode) Lookup(ctx context.Context, name string) (fs.Node, error) {
 	dir.lockMutex()
 	defer dir.unlockMutex()
@@ -185,7 +75,7 @@ func (dir *DirINode) LookupInt(opName string, name string) (fs.Node, error) {
 		return node, nil
 	}
 
-	// Check negative cache before hitting the backend
+	// Check negative cache before hitting the backend.
 	if dir.checkNegativeCache(opName, name) {
 		return nil, syscall.ENOENT
 	}
@@ -198,7 +88,7 @@ func (dir *DirINode) LookupInt(opName string, name string) (fs.Node, error) {
 	return node, nil
 }
 
-// Responds on FUSE request to read directory
+// Responds on FUSE request to read directory.
 func (dir *DirINode) ReadDirAll(ctx context.Context) ([]fuse.Dirent, error) {
 	dir.lockMutex()
 	defer dir.unlockMutex()
@@ -215,14 +105,12 @@ func (dir *DirINode) ReadDirAll(ctx context.Context) ([]fuse.Dirent, error) {
 	entries := make([]fuse.Dirent, 0, len(allAttrs))
 	for _, a := range allAttrs {
 		if dir.FileSystem.IsPathAllowed(dir.AbsolutePathForChild(a.Name)) {
-			// Creating Dirent structure as required by FUSE
 			entries = append(entries, fuse.Dirent{
 				Inode: a.Inode,
 				Name:  a.Name,
-				Type:  a.FuseNodeType()})
-			// Speculatively pre-creating child Dir or File node with cached attributes,
-			// since it's highly likely that we will have Lookup() call for this name
-			// This is the key trick which dramatically speeds up 'ls'
+				Type:  a.FuseNodeType(),
+			})
+			// Speculatively pre-create a child node with cached attributes.
 			dir.addOrUpdateChildInodeAttrs(ReadDir, a.Name, a)
 		}
 	}
@@ -237,9 +125,8 @@ func (dir *DirINode) ReadDirAll(ctx context.Context) ([]fuse.Dirent, error) {
 	return entries, nil
 }
 
-// Performs Stat() query on the backend
+// Performs Stat() query on the backend.
 func (dir *DirINode) statInodeInHopsFS(operation, name string, attrs *Attrs) (fs.Node, error) {
-
 	a, err := dir.FileSystem.getDFSConnector().Stat(path.Join(dir.AbsolutePath(), name))
 	if err != nil {
 		logger.Info("Stat failed on backend", logger.Fields{Operation: operation, Path: path.Join(dir.AbsolutePath(), name), Error: err})
@@ -257,7 +144,7 @@ func (dir *DirINode) statInodeInHopsFS(operation, name string, attrs *Attrs) (fs
 	return inode, nil
 }
 
-// Responds on FUSE Mkdir request
+// Responds on FUSE Mkdir request.
 func (dir *DirINode) Mkdir(ctx context.Context, req *fuse.MkdirRequest) (fs.Node, error) {
 	dir.lockMutex()
 	defer dir.unlockMutex()
@@ -300,7 +187,7 @@ func (dir *DirINode) Mkdir(ctx context.Context, req *fuse.MkdirRequest) (fs.Node
 	return newInode, nil
 }
 
-// Responds on FUSE Create request
+// Responds on FUSE Create request.
 func (dir *DirINode) Create(ctx context.Context, req *fuse.CreateRequest, resp *fuse.CreateResponse) (fs.Node, fs.Handle, error) {
 	dir.lockMutex()
 	defer dir.unlockMutex()
@@ -351,7 +238,7 @@ func (dir *DirINode) Create(ctx context.Context, req *fuse.CreateRequest, resp *
 		Group:     groupName,
 	})
 
-	//update the attributes of the file now
+	// update the attributes of the file now.
 	_, err = dir.statInodeInHopsFS(Create, file.Attrs.Name, &file.Attrs)
 	if err != nil {
 		dir.removeChildInode(Create, req.Name)
@@ -361,7 +248,7 @@ func (dir *DirINode) Create(ctx context.Context, req *fuse.CreateRequest, resp *
 	return file, handle, nil
 }
 
-// Responds on FUSE Remove request
+// Responds on FUSE Remove request.
 func (dir *DirINode) Remove(ctx context.Context, req *fuse.RemoveRequest) error {
 	dir.lockMutex()
 	defer dir.unlockMutex()
@@ -372,7 +259,6 @@ func (dir *DirINode) Remove(ctx context.Context, req *fuse.RemoveRequest) error 
 	err := dir.FileSystem.getDFSConnector().Remove(targetPath)
 	if err == nil {
 		dir.removeChildInode(Remove, req.Name)
-		// Invalidate staging file cache for the removed path
 		if StagingCache != nil {
 			StagingCache.Remove(targetPath)
 		}
@@ -383,7 +269,7 @@ func (dir *DirINode) Remove(ctx context.Context, req *fuse.RemoveRequest) error 
 	return err
 }
 
-// Responds on FUSE Rename request
+// Responds on FUSE Rename request.
 func (srcParent *DirINode) Rename(ctx context.Context, req *fuse.RenameRequest, dstParentDir fs.Node) error {
 	srcParent.lockMutex()
 	defer srcParent.unlockMutex()
@@ -414,26 +300,20 @@ func (srcParent *DirINode) renameInt(operationName, oldName, newName string, dst
 		return err
 	}
 
-	// Transfer staging file cache entry from old path to new path if it exists
 	if StagingCache != nil {
 		StagingCache.Rename(oldPath, newPath)
 	}
 
-	// disconnect src inode
 	if srcInode != nil {
 		srcParent.removeChildInode(Rename, oldName)
 	}
 
-	// disconnect dst inode
 	if dstInode != nil {
 		dstParentDir.(*DirINode).removeChildInode(Rename, newName)
 	}
 
-	// Invalidate negative cache for the new name in the destination directory
 	dstParentDir.(*DirINode).removeNegativeCacheEntry(newName)
 
-	// Upon successful rename, updating in-memory representation of the file entry
-	// file rename
 	if fnode, ok := (srcInode).(*FileINode); ok {
 		logger.Trace("Rename src is file", logger.Fields{Operation: operationName, From: oldPath, To: newPath})
 		fnode.Attrs.Name = newName
@@ -441,7 +321,6 @@ func (srcParent *DirINode) renameInt(operationName, oldName, newName string, dst
 		dstParentDir.(*DirINode).adoptChildInode(Rename, newName, fnode)
 	}
 
-	// dir rename
 	if dnode, ok := (srcInode).(*DirINode); ok {
 		logger.Trace("Rename src is dir", logger.Fields{Operation: operationName, From: oldPath, To: newPath})
 		dnode.Attrs.Name = newName
@@ -453,7 +332,7 @@ func (srcParent *DirINode) renameInt(operationName, oldName, newName string, dst
 	return nil
 }
 
-// Responds on FUSE Rename request
+// Responds on FUSE Rename request.
 func (srcParent *DirINode) Rename2(ctx context.Context, req *fuse.Rename2Request, dstParentDir fs.Node) error {
 	srcParent.lockMutex()
 	defer srcParent.unlockMutex()
@@ -472,7 +351,7 @@ func (srcParent *DirINode) Rename2(ctx context.Context, req *fuse.Rename2Request
 	return srcParent.renameInt(Rename2, req.OldName, req.NewName, dstParentDir, hdfs.RenameOptions(options))
 }
 
-// Responds on FUSE Chmod request
+// Responds on FUSE Chmod request.
 func (dir *DirINode) Setattr(ctx context.Context, req *fuse.SetattrRequest, resp *fuse.SetattrResponse) error {
 	dir.lockMutex()
 	defer dir.unlockMutex()
@@ -502,110 +381,5 @@ func (dir *DirINode) Setattr(ctx context.Context, req *fuse.SetattrRequest, resp
 		return err
 	}
 
-	return nil
-}
-
-// Responds on FUSE request to forget inode
-func (dir *DirINode) Forget() {
-	dir.lockMutex()
-	defer dir.unlockMutex()
-	// inodes are removed on delete and rename operations.
-	// this forget call is redundant and it causes problems.
-	// In the mount point we identify inodes by names.
-	// For example, we remove a file /some/dir/file. Before
-	// the forget call is processed if the user recreates the
-	// file /some/dir/file then processing forget request
-	// would lead to deleting a correct inode
-	// to fix this issue we have to use inode IDs
-
-	// ask parent to remove me from the children list
-	// logger.Debug(fmt.Sprintf("Forget for dir %s", dir.Attrs.Name), nil)
-	// dir.Parent.removeChildInode(Forget, dir.Attrs.Name)
-}
-
-// checkNegativeCache returns true if the name is in the negative cache and not expired.
-// Must NOT hold childrenMutex when calling this.
-func (dir *DirINode) checkNegativeCache(operation, name string) bool {
-	dir.lockChildrenMutex()
-	defer dir.unlockChildrenMutex()
-
-	if dir.negativeCache == nil {
-		return false
-	}
-
-	expiry, ok := dir.negativeCache[name]
-	if !ok {
-		return false
-	}
-
-	if dir.FileSystem.Clock.Now().After(expiry) {
-		delete(dir.negativeCache, name)
-		return false
-	}
-
-	logger.Debug("Negative cache hit", logger.Fields{Operation: operation, Parent: dir.AbsolutePath(), Child: name})
-	return true
-}
-
-// addNegativeCacheEntry adds a name to the negative cache with TTL = CacheAttrsTimeDuration.
-// Must NOT hold childrenMutex when calling this.
-func (dir *DirINode) addNegativeCacheEntry(name string) {
-	dir.lockChildrenMutex()
-	defer dir.unlockChildrenMutex()
-
-	if dir.negativeCache == nil {
-		dir.negativeCache = make(map[string]time.Time)
-	}
-
-	dir.negativeCache[name] = dir.FileSystem.Clock.Now().Add(CacheAttrsTimeDuration)
-}
-
-// removeNegativeCacheEntry removes a name from the negative cache.
-// Must NOT hold childrenMutex when calling this.
-func (dir *DirINode) removeNegativeCacheEntry(name string) {
-	dir.lockChildrenMutex()
-	defer dir.unlockChildrenMutex()
-
-	if dir.negativeCache != nil {
-		delete(dir.negativeCache, name)
-	}
-}
-
-func (dir *DirINode) lockMutex() {
-	dir.dirMutex.Lock()
-}
-
-func (dir *DirINode) unlockMutex() {
-	dir.dirMutex.Unlock()
-}
-
-func (dir *DirINode) lockChildrenMutex() {
-	dir.childrenMutex.Lock()
-}
-
-func (dir *DirINode) unlockChildrenMutex() {
-	dir.childrenMutex.Unlock()
-}
-
-func (dir *DirINode) Symlink(ctx context.Context, req *fuse.SymlinkRequest) (fs.Node, error) {
-	logger.Error("Unsupported Symlink operation.", logger.Fields{Operation: Symlink, Path: dir.AbsolutePath()})
-	return nil, syscall.ENOTSUP
-}
-
-func (dir *DirINode) Readlink(ctx context.Context, req *fuse.ReadlinkRequest) (string, error) {
-	logger.Error("Unsupported Readlink operation.", logger.Fields{Operation: ReadLink, Path: dir.AbsolutePath()})
-	return "", syscall.ENOTSUP
-}
-
-func (dir *DirINode) Link(ctx context.Context, req *fuse.LinkRequest, old fs.Node) (fs.Node, error) {
-	logger.Error("Unsupported Link operation.", logger.Fields{Operation: Link, Path: dir.AbsolutePath()})
-	return nil, syscall.ENOTSUP
-}
-
-// https://libfuse.github.io/doxygen/structfuse__operations.html#abaa2a0bdc9b9955a399ea6973f6f4927
-// Synchronize directory contents
-// All dir operations are first performed on the backend. So no-op
-func (dir *DirINode) Fsync(ctx context.Context, req *fuse.FsyncRequest) error {
-	logger.Info("Fsync called on Dir ", logger.Fields{Operation: Fsync, Path: dir.AbsolutePath()})
 	return nil
 }
